Move grading models and minor changes
'Ungraded' should specify that it talks about AG ungraded, as other gradings might be added in the future.

The position of the grading is now last in the project info table and it only includes the points.
The justification can be found under the corresponding tab
