+++ bp/templates/bp/grading/orga/project_info_misc_grading_desc.html
@@ -1,1 +1,1 @@
-Alle Bewertungen ({{ gradings_count }})
+AG Bewertungen ({{ gradings_count }})
