+++ bp/templates/bp/grading/orga/project_info_table_grade_info.html
@@ -1,10 +1,8 @@
 {% load fontawesome_5 %}
 
 {% if show_aggrade %}
-    <td>Bewertung:</td>
+    <td>AG Bewertung:</td>
     <td>
-        <b>{{ ag_points }} Punkte</b><br>
-        <br>
-        {{ ag_points_justification|linebreaks }}
+        <b>{{ ag_points }} Punkte</b>
     </td>
 {% endif %}
