+++ bp/templates/bp/index/index.html
@@ -13,7 +13,7 @@
                 <p class="card-text">
                     Bewertet: {{ projects_graded_count }}<br><br>
                     <a href="{% url "bp:project_list" %}" class="btn btn-secondary mb-1">Zur Liste</a>
-                    <a href="{% url "bp:project_list_ungraded" %}" class="btn btn-secondary mb-1">Unbewertet</a>
+                    <a href="{% url "bp:project_list_ungraded" %}" class="btn btn-secondary mb-1">AG Unbewertet</a>
                 </p>
             </div>
         </div>
