+++ bp/templates/bp/project_overview.html
@@ -13,7 +13,7 @@
     <div class="float-right mb-4">
         {% if user.is_staff %}
             <a class="btn btn-danger" href="{% url "bp:project_import" %}">Projekte importieren</a>
-            <a class="btn btn-warning" href="{% url "bp:project_list_ungraded" %}">{% fa5_icon "filter" "fas" %} Unbewertet</a>
+            <a class="btn btn-warning" href="{% url "bp:project_list_ungraded" %}">{% fa5_icon "filter" "fas" %} AG Unbewertet</a>
             <a class="btn btn-info" href="{% url "bp:project_export_grades" %}">Punkte für Moodle exportieren</a>
         {% endif %}
     </div>
